# Switch to Alpine based image (BREAKING CHANGE)