--- a/Dockerfile
+++ b/Dockerfile
@@ -2,49 +2,15 @@
 # This Dockerfile requires BuildKit to be enabled, by setting the environment variable
 # DOCKER_BUILDKIT=1
 # see https://docs.docker.com/build/buildkit/#getting-started
-ARG JVM="eclipse-temurin"
-ARG JAVA_VERSION="17"
-# We duplicate the JVM and JAVA_VERSION vars here as renovate will otherwise fail to properly parse
-ARG BASE_IMAGE="eclipse-temurin:17-jre-focal"
-ARG BASE_DIGEST_AMD64="sha256:1a6c81a1b4144fe893e2452a6340d4e3a54ca53356962ed36dea290d9c43400f"
-ARG BASE_DIGEST_ARM64="sha256:78b88c13f8e2a109af3394cafd50377bf1d80ee5d18ee13c1640bf9dfcba1842"
+ARG BASE_IMAGE="alpine:3.18.2"
+ARG BASE_DIGEST_AMD64="sha256:25fad2a32ad1f6f510e528448ae1ec69a28ef81916a004d3629874104f8a7f70"
+ARG BASE_DIGEST_ARM64="sha256:e3bd82196e98898cae9fe7fbfd6e2436530485974dc4fb3b7ddb69134eda2407"
+ARG JDK_IMAGE="azul/zulu-openjdk-alpine:17.0.7-17.42.19"
+ARG JDK_DIGEST_AMD64="sha256:4158692ebc1753d32c7e14b5bb8b96efec8d645c00fbc9341738d85320f5b2de"
+ARG JDK_DIGEST_ARM64="sha256:525168b306464ce0f9e64b85829a1afcad9fe11588bafa61ed0850642a5d6140"
 
 # set to "build" to build zeebe from scratch instead of using a distball
 ARG DIST="distball"
-
-### Init image containing tini and the startup script ###
-FROM ubuntu:jammy as init
-WORKDIR /zeebe
-RUN rm /etc/apt/apt.conf.d/docker-clean
-RUN --mount=type=cache,target=/var/cache/apt,sharing=locked \
-    --mount=type=cache,target=/var/lib/apt,sharing=locked \
-    apt-get -qq update && \
-    apt-get install -y --no-install-recommends tini=0.19.0-1 && \
-    cp /usr/bin/tini .
-COPY --link --chown=1000:0 docker/utils/startup.sh .
-
-### Build zeebe from scratch ###
-FROM maven:3-${JVM}-${JAVA_VERSION} as build
-WORKDIR /zeebe
-ENV MAVEN_OPTS -XX:MaxRAMPercentage=80
-COPY --link . ./
-RUN --mount=type=cache,target=/root/.m2,rw mvn -B -am -pl dist package -T1C -D skipChecks -D skipTests
-RUN mv dist/target/camunda-zeebe .
-
-### Extract zeebe from distball ###
-FROM ubuntu:jammy as distball
-WORKDIR /zeebe
-ARG DISTBALL="dist/target/camunda-zeebe-*.tar.gz"
-COPY --link ${DISTBALL} zeebe.tar.gz
-
-# Remove zbctl from the distribution to reduce CVE related maintenance effort w.r.t to containers
-RUN mkdir camunda-zeebe && \
-    tar xfvz zeebe.tar.gz --strip 1 -C camunda-zeebe && \
-    find . -type f -name 'zbctl*' -delete
-
-### Image containing the zeebe distribution ###
-# hadolint ignore=DL3006
-FROM ${DIST} as dist
 
 ### AMD64 base image ###
 # BASE_DIGEST_AMD64 is defined at the top of the Dockerfile
@@ -60,11 +26,94 @@ FROM ${BASE_IMAGE}@${BASE_DIGEST_ARM64} as base-arm64
 ARG BASE_DIGEST_ARM64
 ARG BASE_DIGEST="${BASE_DIGEST_ARM64}"
 
+### AMD64 JDK image ###
+# JDK_DIGEST_AMD64 is defined at the top of the Dockerfile
+# hadolint ignore=DL3006
+FROM ${JDK_IMAGE}@${JDK_DIGEST_AMD64} as jdk-amd64
+
+### ARM64 JDK image ##
+# JDK_DIGEST_ARM64 is defined at the top of the Dockerfile
+# hadolint ignore=DL3006
+FROM ${JDK_IMAGE}@${JDK_DIGEST_ARM64} as jdk-arm64
+
+### Build custom JRE using the base JDK image
+# hadolint ignore=DL3006
+FROM jdk-${TARGETARCH} as jre-build
+
+# Build a custom JRE which will strip down and compress modules to end up with a smaller Java \
+# distribution than the official Zulu JRE. This JRE will also include useful debugging tools like
+# jcmd, jmod, jps, etc., which take little to no space. Anecdotally, compressing modules add around
+# 10ms to the start up time, which is negligible considering our application takes ~10s to start up.
+# See https://adoptium.net/blog/2021/10/jlink-to-produce-own-runtime/
+# hadolint ignore=DL3018
+RUN --mount=type=cache,target=/var/cache/apk,id=apk,sharing=locked \
+   apk update && apk add binutils && \
+   jlink \
+     --add-modules ALL-MODULE-PATH \
+     --strip-debug \
+     --no-man-pages \
+     --no-header-files \
+     --compress=2 \
+     --output /jre && \
+   rm -rf /jre/lib/src.zip
+
+### Java base ###
+# hadolint ignore=DL3006
+FROM base-${TARGETARCH} as java
+WORKDIR /
+
+# Inherit from previous build stage
+ARG JAVA_HOME=/opt/java/openjdk
+
+# Default to UTF-8 file.encoding
+ENV LANG='en_US.UTF-8' LANGUAGE='en_US:en' LC_ALL='en_US.UTF-8'
+
+# Setup JAVA_HOME and binaries in the path
+ENV JAVA_HOME ${JAVA_HOME}
+ENV PATH $JAVA_HOME/bin:$PATH
+
+# Install missing dependencies and update existing ones to fix any pending security issues
+# hadolint ignore=DL3018
+RUN --mount=type=cache,target=/var/cache/apk,id=apk,sharing=locked \
+    ln -s /var/cache/apk /etc/apk/cache && \
+    apk update && apk upgrade && \
+    apk add java-common java-cacerts musl musl-locales musl-locales-lang tzdata zlib
+
+# Copy JRE from previous build stage
+COPY --from=jre-build /jre ${JAVA_HOME}
+
+### Extract zeebe from distball ###
+# hadolint ignore=DL3006
+FROM base-${TARGETARCH} as distball
+WORKDIR /zeebe
+ARG DISTBALL="dist/target/camunda-zeebe-*.tar.gz"
+COPY --link ${DISTBALL} zeebe.tar.gz
+# Remove zbctl from the distribution to reduce CVE related maintenance effort w.r.t to containers
+RUN mkdir camunda-zeebe && \
+    tar xfvz zeebe.tar.gz --strip 1 -C camunda-zeebe && \
+    find . -type f -name 'zbctl*' -delete
+
+### Build zeebe from scratch ###
+# We use the latest version here to avoid having to track the Java version multiple times
+# hadolint ignore=DL3006,DL3007
+FROM bellsoft/liberica-openjdk-alpine:latest as build
+WORKDIR /zeebe
+ENV MAVEN_OPTS -XX:MaxRAMPercentage=80
+COPY --link . ./
+RUN --mount=type=cache,target=/root/.m2,rw ./mvnw -B -am -pl dist package -T1C -D skipChecks -D skipTests
+RUN mv dist/target/camunda-zeebe . && find . -type f -name 'zbctl*' -delete
+
+### Image containing the zeebe distribution ###
+# This stage is required because we cannot use arguments in the COPY --from statement below, but we
+# can in the FROM statement
+# hadolint ignore=DL3006
+FROM ${DIST} as dist
+
 ### Application Image ###
-# TARGETARCH is provided by buildkit
 # https://docs.docker.com/engine/reference/builder/#automatic-platform-args-in-the-global-scope
 # hadolint ignore=DL3006
-FROM base-${TARGETARCH} as app
+FROM java as app
+
 # leave unset to use the default value at the top of the file
 ARG BASE_IMAGE
 ARG BASE_DIGEST
@@ -103,8 +152,7 @@ ENV ZB_HOME=/usr/local/zeebe \
     ZEEBE_RESTORE=false
 ENV PATH "${ZB_HOME}/bin:${PATH}"
 # Disable RocksDB runtime check for musl, which launches `ldd` as a shell process
-# We know there's no need to check for musl on this image
-ENV ROCKSDB_MUSL_LIBC=false
+ENV ROCKSDB_MUSL_LIBC=true
 
 WORKDIR ${ZB_HOME}
 EXPOSE 26500 26501 26502
@@ -112,7 +160,8 @@ VOLUME /tmp
 VOLUME ${ZB_HOME}/data
 VOLUME ${ZB_HOME}/logs
 
-RUN groupadd -g 1000 zeebe && \
+WORKDIR /zeebe
+RUN addgroup -g 1000 zeebe && \
     adduser -u 1000 zeebe --system --ingroup zeebe && \
     chmod g=u /etc/passwd && \
     # These directories are to be mounted by users, eagerly creating them and setting ownership
@@ -122,8 +171,14 @@ RUN groupadd -g 1000 zeebe && \
     chown -R 1000:0 ${ZB_HOME} && \
     chmod -R 0775 ${ZB_HOME}
 
-COPY --from=init --chown=1000:0 /zeebe/tini ${ZB_HOME}/bin/
-COPY --from=init --chown=1000:0 /zeebe/startup.sh /usr/local/bin/startup.sh
+# Install tini and RocksDB dependencies
+# Additionally, install dependencies required to make sure the image works out of the box with the
+# official Helm chart and SaaS controller
+# hadolint ignore=DL3018
+RUN --mount=type=cache,target=/var/cache/apk,id=apk,sharing=locked \
+    apk update && apk add tini libstdc++ libgcc musl-dev bash
+
+COPY --link --chown=1000:0 docker/utils/startup.sh /usr/local/bin/startup.sh
 COPY --from=dist --chown=1000:0 /zeebe/camunda-zeebe ${ZB_HOME}
 
 ENTRYPOINT ["tini", "--", "/usr/local/bin/startup.sh"]