--- a/Dockerfile
+++ b/Dockerfile
@@ -2,12 +2,12 @@
 # This Dockerfile requires BuildKit to be enabled, by setting the environment variable
 # DOCKER_BUILDKIT=1
 # see https://docs.docker.com/build/buildkit/#getting-started
-ARG BASE_IMAGE="alpine:3.18.2"
-ARG BASE_DIGEST_AMD64="sha256:25fad2a32ad1f6f510e528448ae1ec69a28ef81916a004d3629874104f8a7f70"
-ARG BASE_DIGEST_ARM64="sha256:e3bd82196e98898cae9fe7fbfd6e2436530485974dc4fb3b7ddb69134eda2407"
-ARG JDK_IMAGE="azul/zulu-openjdk-alpine:17.0.7-17.42.19"
-ARG JDK_DIGEST_AMD64="sha256:4158692ebc1753d32c7e14b5bb8b96efec8d645c00fbc9341738d85320f5b2de"
-ARG JDK_DIGEST_ARM64="sha256:525168b306464ce0f9e64b85829a1afcad9fe11588bafa61ed0850642a5d6140"
+ARG BASE_IMAGE="ubuntu:jammy-20230624"
+ARG BASE_DIGEST_AMD64="sha256:b060fffe8e1561c9c3e6dea6db487b900100fc26830b9ea2ec966c151ab4c020"
+ARG BASE_DIGEST_ARM64="sha256:fb4a67ec973b2995214edd101e37a83787b175a16750b372789c8f6314dc20ca"
+ARG JDK_IMAGE="eclipse-temurin:17.0.7_7-jdk-jammy"
+ARG JDK_DIGEST_AMD64="sha256:ab4bbe391a42adc8e590d0c54b3ca7903cbc3b62a3e3b23ac8dce94ebfef6b9e"
+ARG JDK_DIGEST_ARM64="sha256:6af63732bd4bbabe749544e23dac50524b493281208b8b227c7ff87e6d4f2fe9"
 
 # set to "build" to build zeebe from scratch instead of using a distball
 ARG DIST="distball"
@@ -25,6 +25,19 @@ ARG BASE_DIGEST="${BASE_DIGEST_AMD64}"
 FROM ${BASE_IMAGE}@${BASE_DIGEST_ARM64} as base-arm64
 ARG BASE_DIGEST_ARM64
 ARG BASE_DIGEST="${BASE_DIGEST_ARM64}"
+
+### Base image with security updates applied ###
+# hadolint ignore=DL3006
+FROM base-${TARGETARCH} as ubuntu
+
+ARG SECURITY_SOURCES=/etc/apt/security.sources.list
+
+# We only want to apply security patches, so we'll grab only the security related sources and create
+# our own source list. From there, we can upgrade all packages with security vulnerabilities.
+# hadolint ignore=DL4006,DL3008,DL3009
+RUN --mount=type=cache,target=/var/cache/apt \
+  apt-get -yqq update && apt-get -yqq --no-install-recommends upgrade && \
+  apt-get -yqq --no-install-recommends install tini
 
 ### AMD64 JDK image ###
 # JDK_DIGEST_AMD64 is defined at the top of the Dockerfile
@@ -46,9 +59,7 @@ FROM jdk-${TARGETARCH} as jre-build
 # 10ms to the start up time, which is negligible considering our application takes ~10s to start up.
 # See https://adoptium.net/blog/2021/10/jlink-to-produce-own-runtime/
 # hadolint ignore=DL3018
-RUN --mount=type=cache,target=/var/cache/apk,id=apk-jre,sharing=shared \
-   apk update && apk add binutils && \
-   jlink \
+RUN jlink \
      --add-modules ALL-MODULE-PATH \
      --strip-debug \
      --no-man-pages \
@@ -59,7 +70,7 @@ RUN --mount=type=cache,target=/var/cache/apk,id=apk-jre,sharing=shared \
 
 ### Java base ###
 # hadolint ignore=DL3006
-FROM base-${TARGETARCH} as java
+FROM ubuntu as java
 WORKDIR /
 
 # Inherit from previous build stage
@@ -71,17 +82,6 @@ ENV LANG='en_US.UTF-8' LANGUAGE='en_US:en' LC_ALL='en_US.UTF-8'
 # Setup JAVA_HOME and binaries in the path
 ENV JAVA_HOME ${JAVA_HOME}
 ENV PATH $JAVA_HOME/bin:$PATH
-
-# Install missing dependencies and update existing ones to fix any pending security issues
-# hadolint ignore=DL3018
-RUN --mount=type=cache,target=/var/cache/apk,id=apk-java,sharing=shared \
-    ln -s /var/cache/apk /etc/apk/cache && \
-    apk update && apk upgrade && \
-    apk add jemalloc java-common java-cacerts musl musl-locales musl-locales-lang tzdata zlib
-
-# Set jemalloc as the default memory allocator, as it is much more performant than the base malloc
-# implementation in musl
-ENV LD_PRELOAD=/usr/lib/libjemalloc.so.2
 
 # Copy JRE from previous build stage
 COPY --from=jre-build /jre ${JAVA_HOME}
@@ -100,7 +100,7 @@ RUN mkdir camunda-zeebe && \
 ### Build zeebe from scratch ###
 # We use the latest version here to avoid having to track the Java version multiple times
 # hadolint ignore=DL3006,DL3007
-FROM bellsoft/liberica-openjdk-alpine:latest as build
+FROM jdk-${TARGETARCH} as build
 WORKDIR /zeebe
 ENV MAVEN_OPTS -XX:MaxRAMPercentage=80
 COPY --link . ./
@@ -156,7 +156,7 @@ ENV ZB_HOME=/usr/local/zeebe \
     ZEEBE_RESTORE=false
 ENV PATH "${ZB_HOME}/bin:${PATH}"
 # Disable RocksDB runtime check for musl, which launches `ldd` as a shell process
-ENV ROCKSDB_MUSL_LIBC=true
+ENV ROCKSDB_MUSL_LIBC=false
 
 WORKDIR ${ZB_HOME}
 EXPOSE 26500 26501 26502
@@ -165,7 +165,7 @@ VOLUME ${ZB_HOME}/data
 VOLUME ${ZB_HOME}/logs
 
 WORKDIR /zeebe
-RUN addgroup -g 1000 zeebe && \
+RUN groupadd -g 1000 zeebe && \
     adduser -u 1000 zeebe --system --ingroup zeebe && \
     chmod g=u /etc/passwd && \
     # These directories are to be mounted by users, eagerly creating them and setting ownership
@@ -174,13 +174,6 @@ RUN addgroup -g 1000 zeebe && \
     mkdir ${ZB_HOME}/logs && \
     chown -R 1000:0 ${ZB_HOME} && \
     chmod -R 0775 ${ZB_HOME}
-
-# Install tini and RocksDB dependencies
-# Additionally, install dependencies required to make sure the image works out of the box with the
-# official Helm chart and SaaS controller
-# hadolint ignore=DL3018
-RUN --mount=type=cache,target=/var/cache/apk,id=apk-app,sharing=shared \
-    apk update && apk add tini libstdc++ libgcc bash
 
 COPY --link --chown=1000:0 docker/utils/startup.sh /usr/local/bin/startup.sh
 COPY --from=dist --chown=1000:0 /zeebe/camunda-zeebe ${ZB_HOME}