## Description

Switches the base image of our Dockerfile to an Alpine based container. To stay up to date with Alpine (as the Temurin builds can be behind by over a month), we unpack our own JRE into the latest available Alpine image. We also update all the base packages from Alpine to resolve any potential security issues.

> **Note**
> Unfortunately `apk` doesn't let us choose to install only security updates, but considering the base image contains few packages, we consider it safe enough to update everything. 

> **Warning**
> This will break reproducibility of builds however, and should be considered a trade-off.

To stay compatible with the Helm chart and SaaS operator, we install a few missing dependencies in the final image, notably bash. Ideally in the future this will not be necessary and everyone can switch to `sh`.

> **Note** 
> I've already published a sample image under `gcr.io/zeebe-io/zeebe:alpine` if you'd like to try it out.

One thing missing that would need to be done as a follow up, is hook the [Java version updates with Renovate](https://docs.renovatebot.com/modules/datasource/java-version/). I think it won't be possible to update both the version URL and the checksum, but we can at least raise a PR with the tarball URL updated, and manually update the checksum. I'm happy if someone with more Renovate knowledge knows of a better way.

Several Hadolint warnings were ignored:

- Ignore the strict tagging requirements of Hadolint. For the base image, we don't want to pin the sha because it would pin the arch, and that base image is not what we will end up with in the end anyway.
- Ignore pinning the APK package versions, because we actually want to use the latest versions. This breaks reproducibility, but this is on purpose.
- Ignore the `set -o pipefail` since on Alpine we're working with `sh`, not `bash`.
- Ignore that POSIX `sh` does not define process substitution, because it works on Linux `sh` as packaged in BusyBox.

BREAKING CHANGE: this may break existing deployment which assume a Debian or Ubuntu base. Care was taken mostly to be compatible with the existing Helm charts and SaaS k8s operator.

## Related issues

closes #13214 
closes #12959

<!-- Cut-off marker
_All lines under and including the cut-off marker will be removed from the merge commit message_

## Definition of Ready

Please check the items that apply, before requesting a review.

You can find more details about these items in our wiki page about [Pull Requests and Code Reviews](https://github.com/camunda/zeebe/wiki/Pull-Requests-and-Code-Reviews).

* [ ] I've reviewed my own code
* [ ] I've written a clear changelist description
* [ ] I've narrowly scoped my changes
* [ ] I've separated structural from behavioural changes
-->

## Definition of Done

<!-- Please check the items that apply, before merging or (if possible) before requesting a review. -->

_Not all items need to be done depending on the issue and the pull request._

Code changes:
* [ ] The changes are backwards compatibility with previous versions
* [ ] If it fixes a bug then PRs are created to [backport](https://github.com/camunda/zeebe/compare/stable/0.24...main?expand=1&template=backport_template.md&title=[Backport%200.24]) the fix to the last two minor versions. You can trigger a backport by assigning labels (e.g. `backport stable/1.3`) to the PR, in case that fails you need to create backports manually.

Testing:
* [ ] There are unit/integration tests that verify all acceptance criterias of the issue
* [ ] New tests are written to ensure backwards compatibility with further versions
* [ ] The behavior is tested manually
* [ ] The change has been verified by a QA run
* [ ] The impact of the changes is verified by a benchmark

Documentation:
* [ ] The documentation is updated (e.g. BPMN reference, configuration, examples, get-started guides, etc.)
* [ ] If the PR changes how BPMN processes are validated (e.g. support new BPMN element) then the Camunda modeling team should be informed to adjust the BPMN linting.

Other teams:
If the change impacts another team an issue has been created for this team, explaining what they need to do to support this change.
- [ ] [Operate](https://github.com/camunda/operate/issues)
- [ ] [Tasklist](https://github.com/camunda/tasklist/issues)
- [ ] [Web Modeler](https://github.com/camunda/web-modeler/issues)
- [ ] [Desktop Modeler](https://github.com/camunda/camunda-modeler/issues)
- [ ] [Optimize](https://github.com/camunda/camunda-optimize/issues)

Please refer to our [review guidelines](https://github.com/camunda/zeebe/wiki/Pull-Requests-and-Code-Reviews#code-review-guidelines).