@npepinpe 2023-07-04
Seems busybox `ip` behaves a bit differently, and flushing by type does...nothing. :clown_face: