@cmur2 2023-07-04
> One thing missing that would need to be done as a follow up, is hook the [Java version updates with Renovate](https://docs.renovatebot.com/modules/datasource/java-version/). I think it won't be possible to update both the version URL and the checksum, but we can at least raise a PR with the tarball URL updated, and manually update the checksum. I'm happy if someone with more Renovate knowledge knows of a better way.

I feel like I was summoned: in [infra-core](https://github.com/camunda/infra-core/blob/stage/.github/workflows/update-scripts.yml) we introduced a GHA workflow that will be triggered only on Renovate PRs and can execute custom scripts (we use it to regenerate Helm charts) and push those changes to the Renovate branch of the PR. This will be done with a spoofed commit author to ensure that Renovate sees the commit as theirs and continues to update the PR.

I think that solution could also work for you, you'd have to create a script with the Java version update logic incl checksum recalculation (but having such a script could also be useful for humans anyways 😃).