@npepinpe 2023-07-04
I think I got rid of most of the blockers. At this point, the main issue is the following: Temurin does not provide Alpine based binaries for ARM64. This is a work in progress - you can see they have nightlies for ARM64, so hopefully it means this is coming soon. I couldn't find any ETA however on this, so the question is: do we wait for them and postpone our switch to Alpine, or do we switch OpenJDK variant?

If we switch OpenJDK variants, we have four options:

1. [Amazon Corretto](https://aws.amazon.com/corretto/?filtered-posts.sort-by=item.additionalFields.createdDate&filtered-posts.sort-order=desc)
2. [Azul Zulu](https://www.azul.com/downloads/?package=jdk#zulu)
3. [Liberica OpenJDK](https://bell-sw.com/pages/downloads/)
4. [Alpine's OpenJDK package](https://pkgs.alpinelinux.org/package/edge/community/x86_64/openjdk17-jre)

All of these are compliant with Oracle's TCK, and should be functionally equivalent, meaning switching between them should be easy, and mostly based on how easy it is to build the image, track updates, etc., mostly CI/automation stuff.

### Corretto

The main benefit of Corretto is Amazon's promise of long term backwards compatibility. As we're looking to upgrade as soon as possible to the latest LTS, I don't think this is particularly beneficial for us (e.g. long term Java 8 support). The next thing is the number of architectures supported, which is a plus for us, but we still only care about AMD64 and ARM64.

It's also licensed under the GPLv2 with CPE, much like all OpenJDK variants.

Amazon distributes binaries as TAR balls via their website, however without any checksum file as far as I can tell. Considering their UI, we would likely have to depend on Renovate's Java Versions data source for updates.

### Azul Zulu

The main benefit of Zulu is it's often offering bleeding edge features (for example, first JDK to support CRaC, which drastically reduces start up time). Additionally, Azul has been an OpenJDK contributor for a long time now, and they're a big player in the Java world. 

AFAIK their free, OpenJDK variant does not have their custom features (such as their custom, pauseless GC). It's licensed under GPLv2 with CPE as well.

Azul distributes binaries as TAR balls via their websites. Dependency updates would need to either check their website for updates (unclear yet how easy to do), or relying on Renovate's Java Version data source, although it's unclear how closely they follow OpenJDK's versioning.

### Liberica

Liberica is BellSoft's OpenJDK variant, focused on a small size and multiple architecture support. While they aren't as established as Azul, BellSoft are the ones behind Project Portola, so they definitely know what they're doing. Liberica is also the recommended JDK from the Spring project at this point in time.

Liberica JDK is licensed under GPLv2 with CPE as well.

In terms of automation, pre-built binaries are distributed as TAR balls from their website. Dependency updates would need to either check their website for updates (unclear yet how easy to do), or relying on Renovate's Java Version data source, as Liberica is pretty much following OpenJDK versioning. On new update, we can use cmur2's script based method to then fetch the latest Liberica update.

### Alpine's OpenJDK

This is a plain OpenJDK build, built from the GitHub repository. This is a community maintained package however, so relies on open source contributors keeping it up to date.

Looking up patches and so on would boil down to regularly checking with APK if we have new available packages for Java.