@megglos 2023-07-04
From my understanding, license & multiarch is fine for all of them, I try to compare pros and cons on the build integration and maintenance aspects.

🟢  straightforward
🟡  some reasonable effort required
🔴  impossible or high effort

> ### Corretto

Build-integration: 🟢  download via direct link
https://corretto.aws/downloads/resources/17.0.7.7.1/amazon-corretto-17.0.7.7.1-linux-x64.tar.gz
Checksums available: 🟡 yes, for versioned downloads checksums are within the markdown of a release on github 
https://github.com/corretto/corretto-17/releases/tag/17.0.7.7.1
Automate update: 🟡  need to evaluate if it can be covered via renovate with general java versioning, seems like it follows some own versioning though? `17.0.7.7.1`

> ### Azul Zulu

Build-integration: 🟢  download via direct link
https://cdn.azul.com/zulu/bin/zulu17.42.21-ca-crac-jdk17.0.7-linux_x64.tar.gz 
Checksums available: 🔴 so far as I could only find them in a model window in the download page UI
https://docs.azul.com/core/zulu-openjdk/install/alpine-linux#install-from-binary-tar-gz-file
Automate update: 🔴  need to evaluate if it can be covered via renovate with general java versioning, however the additional zulu versioning that is part of the direct links seems to make this even harder

> ### Liberica

Build-integration: 🟢  download via direct link
https://download.bell-sw.com/java/17.0.7+7/bellsoft-jdk17.0.7+7-linux-aarch64.tar.gz
or from their github releases
e.g. https://github.com/bell-sw/Liberica/releases/download/17.0.7%2B7/bellsoft-jdk17.0.7+7-linux-amd64-full.tar.gz
Checksums available: 🟡 yes, all checksums in one txt file on the github release
https://github.com/bell-sw/Liberica/releases/download/17.0.7%2B7/sha1sum.txt
Automate update: 🟡  need to evaluate if it can be covered via renovate with general java versioning


> ### Alpine's OpenJDK

CI-integration: 🟢 
Checksums available: 🟢 
Automate update: 🔴  we have no control an rely on the package maintainers