@npepinpe 2023-07-04
So the Zulu checksums are available, you have to go to the specific download page, where it then offers you the checksum to compare :upside_down_face: Agreed about the versioning however, can't really tell if there's a clear pattern or not.