@npepinpe 2023-07-10
With multiple partitions, we see that the Alpine + mimalloc combo works well. Surprisingly, the normal image with mimalloc performs worse than the normal image itself. Same results with jemalloc. Go figure.

After finding some errors in our benchmarks, I started 3 new benchmarks with a single partition (replicated over 3 nodes) to reduce possible blast radius. Target throughput is 150 PI/s.

- [Baseline achieves 130-140 PI/s, with p50 and p99 latencies of [180ms, 350ms] and [750ms, 2s].](https://grafana.dev.zeebe.io/goto/toIreiC4g?orgId=1)
- [Liberica musl + mimalloc achieves 120-140 PI/s, with p50 and p99 latencies of [200ms, 390ms], [800ms, 1.5s]](https://grafana.dev.zeebe.io/goto/ph-zgZjVR?orgId=1)
- [Alpine musl achieves 120-140 PI/s, with p50 and p99 latencies of [240ms, 400ms], [10s, 10s]]()
- [Baseline with mimalloc achieves 120-130 PI/s, with p50 and p99 latencies of [300ms, 400ms], [10s, 10s]](https://grafana.dev.zeebe.io/goto/SFtbRZC4R?orgId=1)

The performances are all sort of similar. 

I'll try pushing it further to 200 PI/s.